LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/items/row-automatic-minimum-for-auto-expected.txt

XX
XXX
XX
XXX
XXXX
XX
XX
XX
X
X
X
XXX
X
XX
XX
X
X
XXXX
X
XX
XXX
X
X
XX
XXX
XX
XXX
XX
XX XX
XXXXXX X XXX
XX XX XX
X X X X
XX XX
X XXX XX
XXX XXXX
XX XX XX
XXXXX X XXXXX
X XXX XX
XXXXX XXXXX
XX XX
XX XX XX XX
XXX
X XX XXX
X XX X
XXX X
XXX X
XXX X
XX XXX XX XXX XX XXX
XX XXX XX
XX XXX XX X XX
XXXXXX
XXXXXX
XXXXXX
XX
XXX X
XX
XXX X
XX
XXX X
XXX X
XXX X
XXX X

FAIL Check that min-height is honored when sizing auto rows. assert_equals: expected "40px" but got "50px"
FAIL Check that min-height is honored when sizing auto rows and spanning grid items. assert_equals: expected "20px 30px" but got "0px 60px"
FAIL Check the interactions between height and min-height and auto tracks. assert_equals: expected "25px 25px" but got "0px 50px"
PASS Check that borders and paddings are considering when computing min sizes.
